add cast to pacify optimizer

+++ src/time.c
@@ -348,7 +348,7 @@ mrb_time_plus(mrb_state *mrb, mrb_value self)
   mrb_get_args(mrb, "f", &f);
   tm = (struct mrb_time *)mrb_get_datatype(mrb, self, &mrb_time_type);
   if (!tm) return mrb_nil_value();
-  return mrb_time_make(mrb, mrb_obj_class(mrb, self), tm->sec+f, tm->usec, tm->timezone);
+  return mrb_time_make(mrb, mrb_obj_class(mrb, self), (double)tm->sec+f, tm->usec, tm->timezone);
 }
 
 static mrb_value
@@ -370,7 +370,7 @@ mrb_time_minus(mrb_state *mrb, mrb_value self)
   }
   else {
     mrb_get_args(mrb, "f", &f);
-    return mrb_time_make(mrb, mrb_obj_class(mrb, self), tm->sec-f, tm->usec, tm->timezone);
+    return mrb_time_make(mrb, mrb_obj_class(mrb, self), (double)tm->sec-f, tm->usec, tm->timezone);
   }
 }
